chore(core): move last remaining services
