+++ src/bp/core/bpfs/memory-cache.ts
