+++ src/bp/core/dialog/dialog-engine.ts
@@ -1,10 +1,10 @@
 import { FlowNode, IO, Logger } from 'botpress/sdk'
 import { FlowView } from 'common/typings'
 import { createForGlobalHooks } from 'core/app/api'
+import { TYPES } from 'core/app/types'
 import { buildUserKey, converseApiEvents } from 'core/converse'
 import { FlowService } from 'core/dialog'
 import { addErrorToEvent } from 'core/events'
-import { TYPES } from 'core/types'
 import { Hooks, HookService } from 'core/user-code'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
