+++ src/bp/core/dialog/index.ts
@@ -1,11 +1,10 @@
 export * from './state-manager'
-export * from './enums'
+export * from './dialog-engine'
+export * from './decision-engine'
+export * from './janitor'
+export * from './flow/flow-service'
+export * from './skill-service'
+export * from './dialog.inversify'
 
-// TODO: Move files here when references are updated everywhere
-export * from '../services/dialog/dialog-engine'
-export * from '../services/dialog/decision-engine'
-export * from '../services/dialog/janitor'
-export * from '../services/dialog/flow/service'
-export * from '../services/dialog/skill/service'
-export * from '../services/dialog/dialog.inversify'
-export * from '../services/dialog/errors'
+export * from './errors'
+export * from './enums'
