+++ src/bp/core/dialog/instruction/index.ts
