+++ src/bp/core/dialog/instruction/strategy.ts
@@ -1,9 +1,9 @@
 import { IO, Logger } from 'botpress/sdk'
 import { parseActionInstruction } from 'common/action'
 import { ActionServer } from 'common/typings'
+import { TYPES } from 'core/app/types'
 import { CMSService, renderTemplate } from 'core/cms'
 import { EventEngine } from 'core/events'
-import { TYPES } from 'core/types'
 import { ActionService, ActionServersService, VmRunner } from 'core/user-code'
 import { inject, injectable, tagged } from 'inversify'
 import _ from 'lodash'
