+++ src/bp/core/dialog/state-manager.ts
@@ -1,9 +1,9 @@
 import * as sdk from 'botpress/sdk'
+import { TYPES } from 'core/app/types'
 import { BotpressConfig, ConfigProvider } from 'core/config'
 import Database from 'core/database'
 import { JobService } from 'core/distributed'
 import { KeyValueStore } from 'core/kvs'
-import { TYPES } from 'core/types'
 import { ChannelUserRepository } from 'core/users'
 import { inject, injectable, tagged } from 'inversify'
 import { Redis } from 'ioredis'
