+++ src/bp/core/dialog/utils/stubs.ts
