+++ src/bp/core/dialog/utils/validator.ts
